+++ Sources/UnstoppableDomainsResolution/Configurations.swift
@@ -30,15 +30,33 @@ public struct NamingServiceConfig {
     }
 }
 
+public struct UNSConfig {
+    let layer1: NamingServiceConfig
+    let layer2: NamingServiceConfig
+
+    public init(
+        layer1: NamingServiceConfig,
+        layer2: NamingServiceConfig
+    ) {
+        self.layer1 = layer1
+        self.layer2 = layer2
+    }
+}
+
 public struct Configurations {
-    let uns: NamingServiceConfig
+    let uns: UNSConfig
     let zns: NamingServiceConfig
     let ens: NamingServiceConfig
 
     public init(
-        uns: NamingServiceConfig = NamingServiceConfig(
-            providerUrl: "https://mainnet.infura.io/v3/3c25f57353234b1b853e9861050f4817",
-            network: "mainnet"),
+        uns: UNSConfig = UNSConfig(
+            layer1: NamingServiceConfig(
+                providerUrl: "https://mainnet.infura.io/v3/3c25f57353234b1b853e9861050f4817",
+                network: "mainnet"),
+            layer2: NamingServiceConfig(
+                providerUrl: "https://polygon-mumbai.infura.io/v3/3c25f57353234b1b853e9861050f4817",
+                network: "polygon-mumbai")
+        ),
         ens: NamingServiceConfig = NamingServiceConfig(
             providerUrl: "https://mainnet.infura.io/v3/d423cf2499584d7fbe171e33b42cfbee",
             network: "mainnet"),
